+++ modules/apps/portal-vulcan/portal-vulcan-impl/src/main/java/com/liferay/portal/vulcan/internal/batch/engine/action/EntityExtensionImportTaskPostAction.java
@@ -14,31 +14,27 @@
 
 package com.liferay.portal.vulcan.internal.batch.engine.action;
 
-import com.liferay.batch.engine.BatchEngineTaskOperation;
-import com.liferay.batch.engine.action.ImportTaskPreAction;
+import com.liferay.batch.engine.action.ImportTaskPostAction;
 import com.liferay.batch.engine.model.BatchEngineImportTask;
-import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.vulcan.dto.extension.EntityExtensionSupport;
 import com.liferay.portal.vulcan.extension.EntityExtensionHandler;
 import com.liferay.portal.vulcan.extension.ExtensionProviderRegistry;
 import com.liferay.portal.vulcan.extension.util.ExtensionUtil;
 
-import java.io.Serializable;
-
-import java.util.Map;
-
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Carlos Correa
+ * @author Matija Petanjek
  */
-@Component(service = ImportTaskPreAction.class)
-public class ExtendedObjectImportTaskPreAction implements ImportTaskPreAction {
+@Component(service = ImportTaskPostAction.class)
+public class EntityExtensionImportTaskPostAction
+	implements ImportTaskPostAction {
 
 	@Override
-	public void run(BatchEngineImportTask batchEngineImportTask, Object item)
+	public void run(
+			BatchEngineImportTask batchEngineImportTask, Object item,
+			Object persistedItem)
 		throws Exception {
 
 		if (!(item instanceof EntityExtensionSupport)) {
@@ -58,44 +54,10 @@ public class ExtendedObjectImportTaskPreAction implements ImportTaskPreAction {
 		EntityExtensionSupport entityExtensionSupport =
 			(EntityExtensionSupport)item;
 
-		entityExtensionHandler.validate(
+		entityExtensionHandler.setExtendedProperties(
 			batchEngineImportTask.getCompanyId(),
-			entityExtensionSupport.getExtendedProperties(),
-			_isPartialUpdate(batchEngineImportTask));
-	}
-
-	private boolean _isPartialUpdate(
-		BatchEngineImportTask batchEngineImportTask) {
-
-		Map<String, Serializable> parameters =
-			batchEngineImportTask.getParameters();
-
-		if (parameters == null) {
-			return false;
-		}
-
-		String createStrategy = MapUtil.getString(parameters, "createStrategy");
-
-		String updateStrategy = MapUtil.getString(parameters, "updateStrategy");
-
-		BatchEngineTaskOperation batchEngineTaskOperation =
-			BatchEngineTaskOperation.valueOf(
-				batchEngineImportTask.getOperation());
-
-		if ((batchEngineTaskOperation == BatchEngineTaskOperation.CREATE) &&
-			StringUtil.equals(createStrategy, "UPSERT") &&
-			StringUtil.equals(updateStrategy, "PARTIAL_UPDATE")) {
-
-			return true;
-		}
-
-		if ((batchEngineTaskOperation == BatchEngineTaskOperation.UPDATE) &&
-			StringUtil.equals(updateStrategy, "PARTIAL_UPDATE")) {
-
-			return true;
-		}
-
-		return false;
+			batchEngineImportTask.getUserId(), persistedItem,
+			entityExtensionSupport.getExtendedProperties());
 	}
 
 	@Reference
